+++ lamb/types/__init__.py
@@ -3,3 +3,5 @@ __author__ = 'KoNEW'
 
 from .locale import *
 from .device_info import *
+from .intenum import *
+from .image import *